split up cli into modules per command
